minor changes regarding the type is solved for the sync operation in backend

+++ backend/Entities/CommTrn.cs
@@ -9,7 +9,7 @@ namespace COSEC_demo.Entities
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("TrnID")]
-        public long TrnID { get; set; }
+        public decimal TrnID { get; set; } 
 
         [Column("MsgStr")]
         public string? MsgStr { get; set; }
